refactor(e2e): one shared tally for the browser checks (#269)
Three copy-pasted copies of check() and the closing summary become one
module. Adds skip(), which requires a reason and always prints it, and makes
finish() exit non-zero when no check passed — a suite that skipped everything
has proved nothing and must not report success.

Stated as "no check passed" rather than a skip-to-pass ratio deliberately: a
ratio would false-fail the smaller suites on a legitimate two-skip
off-season run. Falsified: a lone skip() exits 1.

Pure refactor otherwise — no suite changes what it asserts. Verified against
production, unchanged counts: rail 36, timezone 16, filter-reveal 34.
