+++ frontend/e2e/verify-rail.mjs
@@ -10,13 +10,9 @@
  */
 import { chromium } from 'playwright';
 import { pinClock } from './fixedNow.mjs';
+import { check, finish } from './results.mjs';
 
 const URL = process.env.URL ?? 'http://localhost:3000/';
-const results = [];
-function check(name, ok, detail) {
-  results.push({ name, ok, detail });
-  console.log(`${ok ? 'PASS' : 'FAIL'}  ${name}${detail ? ` — ${detail}` : ''}`);
-}
 
 const browser = await chromium.launch();
 
@@ -571,10 +567,4 @@ for (const [label, width, zoom] of [['320px', 320, 1], ['200% zoom', 900, 2]]) {
 }
 
 await browser.close();
-
-const failed = results.filter(r => !r.ok);
-console.log(`\n${results.length - failed.length}/${results.length} checks passed`);
-if (failed.length) {
-  console.log('FAILED:\n' + failed.map(f => `  - ${f.name}: ${f.detail ?? ''}`).join('\n'));
-  process.exit(1);
-}
+finish();
